# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger-документация к API каталога автомобилей",
        "title": "cars-service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/cars": {
            "get": {
                "description": "Возвращает список автомобилей, поддерживаются фильтрация и пагинация",
                "produces": [
                    "application/json"
                ],
                "summary": "Получение списка автомобилей",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Регистрационный номер",
                        "name": "regNum",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Марка",
                        "name": "mark",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Модель",
                        "name": "model",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Имя владельца",
                        "name": "ownerName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фамилия владельца",
                        "name": "ownerSurname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Отчество владельца",
                        "name": "ownerPatronymic",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение информации",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат входных данных",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Получает на вход номера автомобилей, выполняет запрос во внешний API для получения недостающих данных и добавляет информацию о новых автомобилях",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Добавление новых автомобилей",
                "parameters": [
                    {
                        "description": "Регистрационные номера",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.addCarsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное добавление информации",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carsResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат входных данных",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carsResponse"
                        }
                    },
                    "409": {
                        "description": "Попытка добавления существующего номера",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carsResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carsResponse"
                        }
                    }
                }
            }
        },
        "/cars/{id}": {
            "get": {
                "description": "Возвращает информацию об автомобиле и его владельце, если автомобиль с указанным id существует",
                "produces": [
                    "application/json"
                ],
                "summary": "Получение информации об автомобиле по id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id автомобиля",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение информации",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат входных данных",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    },
                    "404": {
                        "description": "Автомобиль с указанным id не найден",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Изменяет информацию о существующем автомобиле",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Изменение информации об автомобиле",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id автомобиля",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новая информация об автомобиле",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.carData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное изменение информации",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат входных данных",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    },
                    "404": {
                        "description": "Автомобиль с указанным id не найден",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    },
                    "409": {
                        "description": "Попытка добавления существующего номера",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет информацию об автомобиле по его id",
                "produces": [
                    "application/json"
                ],
                "summary": "Удаление информации об автомобиле",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id автомобиля",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление информации",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат входных данных",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    },
                    "404": {
                        "description": "Автомобиль с указанным id не найден",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка на стороне сервера",
                        "schema": {
                            "$ref": "#/definitions/httpserver.carResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httpserver.addCarsRequest": {
            "type": "object",
            "properties": {
                "regNums": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "httpserver.carData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "mark": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/httpserver.ownerData"
                },
                "regNum": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "httpserver.carResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/httpserver.carData"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "httpserver.carsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpserver.carData"
                    }
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "httpserver.ownerData": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        }
    }
}